modularize to server, move example code to example/ folder for now
