+++ .gitignore
@@ -1,2 +1,4 @@
 node_modules
+example/bundle.js
 bundle.js
+example/npm-debug.log
