+++ example/cli.js
@@ -1,33 +1,22 @@
 #!/usr/bin/env node
 
-var stats = require('./')
+var stats = require('./server.js')
 var http = require('http')
+var memdb = require('memdb')
 var swarm = require('hyperdrive-archive-swarm')
 var minimist = require('minimist')
 
 var argv = minimist(process.argv.slice(2), {
-  alias: {port: 'p', 'hyperdrive': 'd', wait: 'w'},
-  boolean: ['hyperdrive']
+  alias: {port: 'p', wait: 'w'}
 })
 
-var hypercore = require('hypercore')
 var hyperdrive = require('hyperdrive')
-var memdb = require('memdb')
-
-var core = hypercore(memdb())
-var f = core.createFeed(argv._[0])
-var archive = null
-
-if (argv.hyperdrive) {
-  archive = hyperdrive(memdb()).createArchive(argv._[0])
-}
-
-var server = http.createServer(stats(archive || f))
+var archive = hyperdrive(memdb()).createArchive(argv._[0])
+var server = http.createServer(stats(archive))
 
 server.on('listening', function () {
-  console.log('Feed: ' + f.key.toString('hex'))
+  console.log('Feed: ' + archive.key.toString('hex'))
   console.log('Stats listening on port ' + server.address().port)
-
   if (argv.wait) setTimeout(join, Number(argv.wait))
   else join()
 })
@@ -38,5 +27,6 @@ server.once('error', function () {
 })
 
 function join () {
-  swarm(archive || f)
+  console.log('joining swarm')
+  swarm(archive)
 }
