+++ package.json
@@ -11,7 +11,7 @@
     "speedometer": "^1.0.0"
   },
   "scripts": {
-    "install": "browserify app.js -o bundle.js"
+    "build": "browserify example/app.js > example/bundle.js"
   },
   "repository": {
     "type": "git",
@@ -22,5 +22,11 @@
   "bugs": {
     "url": "https://github.com/mafintosh/hypercore-stats-ui/issues"
   },
-  "homepage": "https://github.com/mafintosh/hypercore-stats-ui"
+  "homepage": "https://github.com/mafintosh/hypercore-stats-ui",
+  "devDependencies": {
+    "hypercore-stats-server": "^1.0.0",
+    "hyperdrive-archive-swarm": "^3.1.0",
+    "memdb": "^1.3.1",
+    "minimist": "^1.2.0"
+  }
 }
